Custom Output Data from Raven

A dataset formatted to the xts package, read in by the rvn_custom_read function. The dataset contains average SNOW for each HRU in the Nith river model, available for download in the Raven Tutorials (linked below).

Note that this data set cannot be used with rvn_custom_output_plot as the file name information is not available in this data format. Please refer to the example in the plotting function to use the sample data file directly, which includes the filename information.

The Nith River model can be downloaded from the Raven Tutorials (tutorial #2) http://www.civil.uwaterloo.ca/jrcraig/Raven/Downloads.html

rvn_custom_data

Format

A data frame with 730 rows, containg data for 32 HRUs from 2002-10-01 to 2004-09-29

See also

Examples

# Preview data
head(rvn_custom_data)
#>                     1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20 21 22 23
#> 2002-10-01 04:00:00 0 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> 2002-10-02 04:00:00 0 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> 2002-10-03 04:00:00 0 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> 2002-10-04 04:00:00 0 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> 2002-10-05 04:00:00 0 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> 2002-10-06 04:00:00 0 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#>                     24 25 26 27 28 29 30 31 32
#> 2002-10-01 04:00:00  0  0  0  0  0  0  0  0  0
#> 2002-10-02 04:00:00  0  0  0  0  0  0  0  0  0
#> 2002-10-03 04:00:00  0  0  0  0  0  0  0  0  0
#> 2002-10-04 04:00:00  0  0  0  0  0  0  0  0  0
#> 2002-10-05 04:00:00  0  0  0  0  0  0  0  0  0
#> 2002-10-06 04:00:00  0  0  0  0  0  0  0  0  0