Forcings Data from Raven

A dataset formatted to the xts package, read in by the forcings.read function. The dataset contains the typical columns from the Raven outputted ForcingFunctions.csv file, available for download in the Raven Tutorials (linked below).

rvn_forcing_data

Format

rvn_forcing_data is a data frame with two object

forcings

various forcing functions and related output from Raven model

units

units associated with each variable in forcings

rvn_forcing_data$forcings is an xts (time series) object with 731 rows and 21 variables, containing data from 2002-10-01 to 2004-09-30. The details of each forcing function can be found in the Raven Manual

day_angle

rain

snow

temp

temp_daily_min

temp_daily_max

temp_daily_ave

temp_monthly_min

temp_monthly_max

air_dens

air_pres

rel_hum

cloud_cover

ET_radiation

SW_radiation

LW_radiation

wind_vel

PET

OW_PET

daily_correction

potential_melt

The Nith River model can be downloaded from the Raven Tutorials (tutorial #2) http://www.civil.uwaterloo.ca/jrcraig/Raven/Downloads.html

See also