Plot Raven reservoir/lake stage time series using dygraph

rvn_res_dygraph plots modeled vs observed stage plots when supplied with reservoir stage data structure read using rvn_res_read

rvn_res_dygraph(resdata, timezone = "UTC", basins = "", figheight = 400)

Arguments

resdata

reservoir stage time series data structure generated by rvn_res_read routine

timezone

data timezone; defaults to UTC

basins

list of subbasin names from reservoir file. Each subbasin creates separate dygraph plots

figheight

height of figure, in pixels

Value

a list of plot handles to dygraph plots

See also

Examples

# read in Raven Reservoir Stages file
ff <- system.file("extdata","ReservoirStages.csv", package="RavenR")
resdata <- rvn_res_read(ff)

# view contents for all subbasins as dyGraph
dyplots <- rvn_res_dygraph(resdata)
dyplots[[1]]

dyplots[[2]]