Generate Hydrological process connections list

This routine reads in a hydrologic process list from rvn_rvi_read and generates the list of hydrologic process connections.

rvn_rvi_connections(
  rvi,
  ProcConDataFile = system.file("extdata", "RavenProcessConnections.dat", package =
    "RavenR")
)

Arguments

rvi

data object generated from the rvn_rvi_read routine

ProcConDataFile

(optional) path to RavenProcesConnections.dat file

Value

Returns a list with two items:

connections

a a dataframe of all of the process connections Includes the following data columns: process type, algorithm, 'from' compartment, 'to' compartment, and conditional

AliasTable

a table of aliases (unchanged from the supplied rvi$AliasTable)

Details

Relies on a valid and up-to-date RavenProcessConnections.dat file. This file is provided with the RavenR package, but may be overridden by a more recent file if provided manually.

See also

Author

James R. Craig, University of Waterloo

Examples

rvi <- rvn_rvi_read(system.file("extdata","Nith.rvi", package="RavenR"))

rvi_conn <- rvn_rvi_connections(rvi)
head(rvi_conn$connections)
#>             Algorithm   ProcessType         From           To Conditional Note
#> 101 FREEZE_DEGREE_DAY  SnowRefreeze     SNOW_LIQ         SNOW                 
#> 1        PRECIP_RAVEN Precipitation ATMOS_PRECIP PONDED_WATER                 
#> 2        PRECIP_RAVEN Precipitation ATMOS_PRECIP   DEPRESSION                 
#> 3        PRECIP_RAVEN Precipitation ATMOS_PRECIP      WETLAND                 
#> 4        PRECIP_RAVEN Precipitation ATMOS_PRECIP       CANOPY                 
#> 5        PRECIP_RAVEN Precipitation ATMOS_PRECIP  CANOPY_SNOW                 
head(rvi_conn$AliasTable)
#>            alias basename
#> 1 FAST_RESERVOIR  SOIL[1]
#> 2 SLOW_RESERVOIR  SOIL[2]