Retrieve Parameter Information from RVI file Algorithms

Reads the processed rvi object and returns a data frame of parameter information for parameters associated with processes in the rvi object.

rvn_rvi_getparams(
  rvi,
  RavenAlgParamsFile = system.file("extdata", "RavenAlgParams.dat", package = "RavenR"),
  RavenParamsFile = system.file("extdata", "RavenParameters.dat", package = "RavenR")
)

Arguments

rvi

data object generated from the rvn_rvi_read routine

RavenAlgParamsFile

(optional) path to RavenAlgParams.dat file

RavenParamsFile

(optional) path to RavenParameters.dat file

Value

Returns a data frame with parameter information containing the parameter name, parameter class, units, auto (whether the parameter may be autogenerated within Raven), default, min, and max values. Note that the flags 'xxx', -9999, and 9999 are used as missing values within the parameter data frame.

Details

Uses the Raven database files in the /extdata folder to (1) associate parameters with particular algorithms, and (2) subset the database parameter list based on the information in the rvi file.

These files are stored with the RavenR package and retrieved with this function by default, but a separate link may be provided to a modified file if desired. Note that the database files held in the RavenR package are unofficial copies of those in the official Raven SVN, and any discrepancies should defer to the Raven SVN versions.

Examples

# sample workflow of rvn_rvi_read
rvi <- system.file("extdata","Nith.rvi", package="RavenR") %>%
rvn_rvi_read(.)

# get data frame of parameters related to processes in rvi file
rvn_rvi_getparams(rvi)
#>                param class_type     units  auto default    min     max
#> 1    FOREST_COVERAGE       LULT    [0..1] False       0      0     1.0
#> 5   IMPERMEABLE_FRAC       LULT    [0..1] False     0.0      0     1.0
#> 13   REFREEZE_FACTOR       LULT [mm/dC/d] False    5.04    0.0    10.0
#> 37          POROSITY       SOIL    [0..1]  True     0.4    0.0     1.0
#> 42    PET_CORRECTION       SOIL    [0..1]  True     1.0    0.0     1.0
#> 43    FIELD_CAPACITY       SOIL    [0..1]  True     0.0    0.0     1.0
#> 44          SAT_WILT       SOIL    [0..1]  True     0.0    0.0     1.0
#> 48          HBV_BETA       SOIL       [-] False       1  -9999 -9999.0
#> 49     MAX_PERC_RATE       SOIL    [mm/d] False    0.05  -9999 -9999.0
#> 55 MAX_CAP_RISE_RATE       SOIL    [mm/d] False       0  -9999 -9999.0
#> 56    BASEFLOW_COEFF       SOIL     [1/d] False     0.1 0.0001     0.5
#> 57        BASEFLOW_N       SOIL       [-] False     2.0    0.5     5.0
#> 71          SNOW_SWI     GLOBAL    [0..1]  True    0.05    0.0     0.5