Rvt Mappings Data

A list of lists that has information for use in the RavenR rvt-related files.

Additional information on Raven variables can be found in the Raven User's Manual, available from http://raven.uwaterloo.ca/Downloads.html

Format

rvn_rvt_mappings_data is a list with four lists.

rvt_mapping

list: data formats for all rvt file types

rvt_data_type_mapping

list: data types permitted in Raven, and associated units

rvn_met_raven_mapping

list: meteorological forcing functions permitted in Raven, and associated units

rvt_met_mapping_weathercan

list: mapping of weathercan variable names to Raven variables

See also