Watershed Storage Data from Raven

A dataset formatted to the xts package, read in by the watershed.read function. The dataset contains the typical columns from the Raven outputted WatershedStorage.csv file, available for download in the Raven Tutorials (linked below).

rvn_watershed_data

Format

rvn_watershed_data is a data frame with two object

watershed.storage

various storage variable states outputted from the Raven model

units

units associated with each variable in watershed.storage

rvn_watershed_data$watershed.storage is an xts (time series) object with 731 rows and 19 variables, containing data from 2002-10-01 to 2004-09-30. The details of each watershed storage state can be found in the Raven Manual

rainfall

snowfall

Channel.Storage

Reservoir.Storage

Rivulet.Storage

Surface.Water

Cum..Losses.to.Atmosphere..mm.

Ponded.Water

Soil.Water.0

Soil.Water.1

Soil.Water.2

Snow.Melt..Liquid..mm.

Snow

Canopy

Canopy.Snow

Total

Cum..Inputs..mm.

Cum..Outflow..mm.

MB.Error

The Nith River model can be downloaded from the Raven Tutorials (tutorial #2) http://www.civil.uwaterloo.ca/jrcraig/Raven/Downloads.html

See also

Examples

# View data
head(rvn_watershed_data$watershed.storage)
#>                     rainfall snowfall Channel.Storage Reservoir.Storage
#> 2002-10-01 04:00:00       NA       NA         1.94009                 0
#> 2002-10-02 04:00:00 3.468690        0         2.79613                 0
#> 2002-10-03 04:00:00 1.189180        0         3.63304                 0
#> 2002-10-04 04:00:00 2.083260        0         3.27268                 0
#> 2002-10-05 04:00:00 6.474310        0         2.42239                 0
#> 2002-10-06 04:00:00 0.125591        0         1.70140                 0
#>                     Rivulet.Storage Surface.Water
#> 2002-10-01 04:00:00        1.056200             0
#> 2002-10-02 04:00:00        2.330950             0
#> 2002-10-03 04:00:00        1.342180             0
#> 2002-10-04 04:00:00        0.940287             0
#> 2002-10-05 04:00:00        0.911627             0
#> 2002-10-06 04:00:00        0.853097             0
#>                     Cum..Losses.to.Atmosphere..mm. Ponded.Water Soil.Water.0
#> 2002-10-01 04:00:00                        0.00000  0.00000e+00     10.00000
#> 2002-10-02 04:00:00                        5.65641  0.00000e+00      7.05032
#> 2002-10-03 04:00:00                       10.22230  1.01511e-17      4.15461
#> 2002-10-04 04:00:00                       12.83760 -9.44248e-18      4.12491
#> 2002-10-05 04:00:00                       18.72080 -7.13371e-17      3.79521
#> 2002-10-06 04:00:00                       21.40350  3.30391e-18      2.12754
#>                     Soil.Water.1 Soil.Water.2 Snow.Melt..Liquid..mm. Snow
#> 2002-10-01 04:00:00   10.0000000      10.0000                      0    0
#> 2002-10-02 04:00:00    6.5490500      10.9424                      0    0
#> 2002-10-03 04:00:00    3.9050700      11.8432                      0    0
#> 2002-10-04 04:00:00    1.8214000      12.7045                      0    0
#> 2002-10-05 04:00:00    1.1999900      13.5278                      0    0
#> 2002-10-06 04:00:00    0.0917699      13.1340                      0    0
#>                     Canopy Canopy.Snow   Total Cum..Inputs..mm.
#> 2002-10-01 04:00:00      0           0 32.9963          0.00000
#> 2002-10-02 04:00:00      0           0 35.3252          3.46869
#> 2002-10-03 04:00:00      0           0 35.1004          4.65787
#> 2002-10-04 04:00:00      0           0 35.7013          6.74113
#> 2002-10-05 04:00:00      0           0 40.5779         13.21540
#> 2002-10-06 04:00:00      0           0 39.3113         13.34100
#>                     Cum..Outflow..mm.     MB.Error
#> 2002-10-01 04:00:00           0.00000  0.00000e+00
#> 2002-10-02 04:00:00           1.13976 -6.21725e-15
#> 2002-10-03 04:00:00           2.55374  1.77636e-15
#> 2002-10-04 04:00:00           4.03608  5.32907e-15
#> 2002-10-05 04:00:00           5.63382  7.10543e-15
#> 2002-10-06 04:00:00           7.02598  1.77636e-15
# Also has units
rvn_watershed_data$units
#>  [1] "mm/day"    "mm.d.SWE." "mm."       "mm."       "mm."       "mm."      
#>  [7] NA          "mm."       "mm."       "mm."       "mm."       NA         
#> [13] "mm."       "mm."       "mm."       "mm."       NA          NA         
#> [19] "mm."